Module 3_Week 3/Day 1_Flask/requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
flask-marshmallow==1.3.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
marshmallow==3.26.0
marshmallow-sqlalchemy==1.4.0
mysql-connector-python==9.2.0
packaging==24.2
python-dotenv==1.0.1
SQLAlchemy==2.0.37
typing_extensions==4.12.2
Werkzeug==3.1.3
